Updated header and footer and added functions.php in artistry-showcase theme
This commit includes the creation of a new functions.php file with all the necessary setup and initialization for the artistry-showcase theme. The header.php file was also updated to retrieve the page title with wp_title() function and the navigation menu was enhanced. Lastly, changes were made in the footer.php to include a navigational menu.